moved changes to main file
